Heap of cleanup
* Removed test, coverage, lint support
* Remove sass support
* Updated linting & fixed errors
* Consumed typescript-log for log interface
* Removed husky and hooks
* Standardised build to include cjs, esm & es2015
* Removed teamcity reporter

+++ test/bin/watch.test.ts
@@ -1,43 +1,57 @@
-import watch from '../../lib/bin/watch'
 import path from 'path'
+import { consoleLogger } from 'typescript-log'
+
+import watch from '../../lib/bin/watch'
+import { getConfig } from '../../lib/runtime/config/config'
 import { waitForConnection } from '../../lib/runtime/util/network'
 import { getTestPort } from '../test-helpers'
-import { getConfig } from '../../lib/runtime/config/config'
-
-// Increase test timeout because builds might take a while
-import { createConsoleLogger } from '../../lib/runtime/universal'
-;(jasmine as any).DEFAULT_TIMEOUT_INTERVAL = 60000
 
 const testProjectDir = path.join(process.cwd(), './test/test-project')
-const log = createConsoleLogger()
+const log = consoleLogger()
 const buildConfig = getConfig(log, testProjectDir)
 
 describe('bin/watch', () => {
+    jest.setTimeout(60000)
+
     it('will watch', async () => {
         const port = await getTestPort()
         buildConfig.DEV_SERVER_PORT = port
-        const childProcess: any = await watch(log, buildConfig, {
-            NODE_ENV: 'production',
-            PROJECT_DIR: testProjectDir,
-        })
-        await waitForConnection(port)
-        childProcess.kill()
+        let childProcess: any
+
+        try {
+            childProcess = await watch(log, buildConfig, {
+                NODE_ENV: 'production',
+                PROJECT_DIR: testProjectDir,
+            })
+            await waitForConnection(port)
+        } catch (err) {
+            if (childProcess) {
+                childProcess.kill()
+            }
+        }
     })
 
     // can't test in TypeScript land because it requires the internal server in JavaScript
     it('will watch the client', async () => {
         const port = await getTestPort()
         buildConfig.DEV_SERVER_PORT = port
-        const childProcess: any = await watch(
-            log,
-            buildConfig,
-            {
-                NODE_ENV: 'production',
-                PROJECT_DIR: testProjectDir,
-            },
-            'client',
-        )
-        await waitForConnection(port)
-        childProcess.kill()
+        let childProcess: any
+
+        try {
+            childProcess = await watch(
+                log,
+                buildConfig,
+                {
+                    NODE_ENV: 'production',
+                    PROJECT_DIR: testProjectDir,
+                },
+                'client',
+            )
+            await waitForConnection(port)
+        } catch (err) {
+            if (childProcess) {
+                childProcess.kill()
+            }
+        }
     })
 })
